Move enable_amsc_x_api_key to shared utils file
